requirements.txt
torch>=1.9.0
transformers>=4.15.0
faiss-cpu>=1.7.2
numpy>=1.19.5
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
